requirements.txt
asgiref==3.4.1
autopep8==1.6.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==38.0.4
cycler==0.11.0
defusedxml==0.7.1
diff-match-patch==20200713
dj-database-url==0.5.0
Django==3.2.13
django-allauth==0.51.0
django-cors-headers==3.13.0
django-environ==0.8.1
django-filter==22.1
django-import-export==3.0.2
django-sslserver==0.22
django-widget-tweaks==1.4.12
djangorestframework==3.13.1
et-xmlfile==1.1.0
fonttools==4.38.0
gunicorn==20.1.0
idna==3.4
kiwisolver==1.4.4
MarkupPy==1.14
oauthlib==3.2.2
odfpy==1.4.1
openpyxl==3.0.10
packaging==22.0
Pillow==9.3.0
pycodestyle==2.8.0
pycparser==2.21
PyJWT==2.6.0
pyparsing==3.0.9
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2021.3
PyYAML==6.0
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.2
tablib==3.3.0
toml==0.10.2
tzdata==2022.7
urllib3==1.26.13
whitenoise==5.3.0
xlrd==2.0.1
xlwt==1.3.0
